# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Forge Core API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: projectId
          in: query
          required: false
          schema: { type: integer }
        - name: status
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update task
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete task
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /decisions:
    get:
      operationId: listDecisions
      tags: [tasks]
      summary: List decisions
      parameters:
        - name: projectId
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: Decisions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Decision"
    post:
      operationId: createDecision
      tags: [tasks]
      summary: Record decision
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecisionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Decision"
  /risks:
    get:
      operationId: listRisks
      tags: [tasks]
      summary: List risks
      parameters:
        - name: projectId
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: Risks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Risk"
    post:
      operationId: createRisk
      tags: [tasks]
      summary: Create risk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Risk"
  /risks/{id}:
    patch:
      operationId: updateRisk
      tags: [tasks]
      summary: Update risk
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Risk"
...
  /sandboxes:
    get:
      operationId: listSandboxes
      tags: [sandboxes]
      summary: List sandboxes
      responses:
        "200":
          description: Sandboxes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sandbox"
    post:
      operationId: createSandbox
      tags: [sandboxes]
      summary: Create sandbox workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SandboxInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sandbox"
  /sandboxes/{id}:
    get:
      operationId: getSandbox
      tags: [sandboxes]
      summary: Get sandbox with files
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Sandbox
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SandboxDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSandbox
      tags: [sandboxes]
      summary: Delete sandbox completely
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /sandboxes/{id}/files:
    post:
      operationId: upsertSandboxFile
      tags: [sandboxes]
      summary: Create or update a file inside a sandbox
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SandboxFileInput"
      responses:
        "200":
          description: File saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SandboxFile"
        "403":
          description: Blocked (attempt to write core/protected path)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sandboxes/{id}/files/{fileId}:
    delete:
      operationId: deleteSandboxFile
      tags: [sandboxes]
      summary: Delete a sandbox file
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: fileId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Task:
      type: object
      required: [id, title, status, risk, ownerAgent, createdAt]
      properties:
        id: { type: integer }
        projectId: { type: ["integer", "null"] }
        parentTaskId: { type: ["integer", "null"] }
        title: { type: string }
        goal: { type: ["string", "null"] }
        scope: { type: ["string", "null"] }
        risk: { type: string }
        ownerAgent: { type: string }
        status:
          type: string
          enum: [draft, planned, active, blocked, testing, review, approved, rejected, done]
        acceptanceCriteria: { type: ["string", "null"] }
        blockedReason: { type: ["string", "null"] }
        source: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: ["string", "null"] }
    TaskInput:
      type: object
      required: [title]
      properties:
        projectId: { type: integer }
        parentTaskId: { type: integer }
        title: { type: string, minLength: 1 }
        goal: { type: string }
        scope: { type: string }
        risk: { type: string }
        ownerAgent: { type: string }
        status:
          type: string
          enum: [draft, planned, active, blocked, testing, review, approved, rejected, done]
        acceptanceCriteria: { type: string }
        source: { type: string }
    TaskUpdate:
      type: object
      properties:
        projectId: { type: integer }
        title: { type: string, minLength: 1 }
        goal: { type: string }
        scope: { type: string }
        risk: { type: string }
        ownerAgent: { type: string }
        status:
          type: string
          enum: [draft, planned, active, blocked, testing, review, approved, rejected, done]
        acceptanceCriteria: { type: string }
        blockedReason: { type: string }
    Decision:
      type: object
      required: [id, title, madeBy, createdAt]
      properties:
        id: { type: integer }
        projectId: { type: ["integer", "null"] }
        taskId: { type: ["integer", "null"] }
        title: { type: string }
        rationale: { type: ["string", "null"] }
        madeBy: { type: string }
        createdAt: { type: string }
    DecisionInput:
      type: object
      required: [title]
      properties:
        projectId: { type: integer }
        taskId: { type: integer }
        title: { type: string, minLength: 1 }
        rationale: { type: string }
        madeBy: { type: string }
    Risk:
      type: object
      required: [id, title, severity, status]
      properties:
        id: { type: integer }
        projectId: { type: ["integer", "null"] }
        title: { type: string }
        severity: { type: string }
        mitigation: { type: ["string", "null"] }
        status: { type: string }
    RiskInput:
      type: object
      required: [title]
      properties:
        projectId: { type: integer }
        title: { type: string, minLength: 1 }
        severity: { type: string }
        mitigation: { type: string }
        status: { type: string }
    RiskUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        severity: { type: string }
        mitigation: { type: string }
        status: { type: string }
...
    Sandbox:
      type: object
      required: [id, name, status, testStatus, createdAt]
      properties:
        id: { type: integer }
        moduleId: { type: ["integer", "null"] }
        name: { type: string }
        purpose: { type: ["string", "null"] }
        status: { type: string }
        testStatus: { type: string }
        storagePath: { type: ["string", "null"] }
        createdAt: { type: string }
    SandboxDetail:
      type: object
      required: [sandbox, files]
      properties:
        sandbox:
          $ref: "#/components/schemas/Sandbox"
        files:
          type: array
          items:
            $ref: "#/components/schemas/SandboxFile"
    SandboxInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        moduleId: { type: integer }
        purpose: { type: string }
    SandboxFile:
      type: object
      required: [id, sandboxId, path, content, updatedAt]
      properties:
        id: { type: integer }
        sandboxId: { type: integer }
        path: { type: string }
        content: { type: string }
        updatedAt: { type: string }
    SandboxFileInput:
      type: object
      required: [path, content]
      properties:
        path: { type: string, minLength: 1 }
        content: { type: string }
...